requirements.txt
marshmallow==2.14.0
sqlalchemy==1.1.15
psycopg2==2.7.4
pyyaml==3.12
alembic==0.8.8
uvloop==0.9.1
asyncpg==0.16.0
aiohttp==3.3.2
aioredis==1.1.0